backend/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Database & Cache
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0

# Environment variables
python-decouple==3.8

# Development & Testing
pytest==7.4.3
pytest-django==4.6.0
coverage==7.3.2
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Code quality
flake8==6.1.0
black==23.9.1
isort==5.12.0

